Stage 2: Reimplement JavaScript logic in C++
Use a variant of the p-impl idiom by having an "impl" property
on the root object of components to hold a QtObject C++ component
that can easily be accessed from QML, and also from C++ through
a QuickImplPointer template member.
Because the QML engine doesn't make the QML component's context
available to the root object, we had to place the C++ object as
a child and not the root object itself.

The reimplementation is a direct translation to C++, the code has
been modified as little as possible to be able to measure the
language factor.

Change-Id: I267fcd4b5d9a8cc1fdb218e44e3fe1a27e84638d